add generic code
